# [libc++] Optimize most of the __tree search algorithms