--- a/libcxx/test/std/input.output/iostreams.base/fpos/fpos.operations/fpos.pass.cpp
+++ b/libcxx/test/std/input.output/iostreams.base/fpos/fpos.operations/fpos.pass.cpp
@@ -14,6 +14,7 @@
 #include <cassert>
 #include <ios>
 #include <type_traits>
+#include <utility>
 
 #include "test_macros.h"