--- a/libcxx/test/support/any_helpers.h
+++ b/libcxx/test/support/any_helpers.h
@@ -9,9 +9,10 @@
 #ifndef ANY_HELPERS_H
 #define ANY_HELPERS_H
 
+#include <cassert>
 #include <typeinfo>
 #include <type_traits>
-#include <cassert>
+#include <utility>
 
 namespace std { namespace experimental {} }